requirements.txt
python-dotenv
google-genai
google-api-python-client
google-auth
google-auth-oauthlib
requests
